Switch to using custom conditions for local imports
Rather than require building type definitions with
sourcemaps to enable jump-to-definition between
packages in the monorepo, switch to using custom
conditions for local imports, to resolve the
typescript modules directly.

As a result, it's not necessary to run a build
to have changes reflected in editors,
or to even run the example projects,
as with tsx they can be imported directly
inside node.

This follows the guide outlined here:
https://colinhacks.com/essays/live-types-typescript-monorepo