Switch to default tsup definitions output
Given that local development has now switched to
using custom conditions, it's no longer necessary
for definitions modules to have sourcmaps,
and in fact most usage of these sourcemaps won't
work as the original source files will not be
present at that time.

So this change introduced in #31 can be reverted.