Refactor to make core components an extension
Rather than have the core components built in to the protocol & toolkit
packages, namespace all existing components under a new toolkit extension
called core.

This means that the core components are a concrete example of how the
toolkit can be extended with additional components for custom use-cases,
and also proves that the extensibility works.