Refactor events to always use props (#84)
* Address esbuild warnings of export ordering

esbuild was complaining about the order in which the types exports were
defined relative to require and import. These have now been reordered to
prevent the warnings.

* Refactor events to always use props

To pave the way for allowing for easier creation of custom components,
rather than having hard-coded event listener updates in react-toolkit,
just expose the listener props as properties in the `@arcanejs/toolkit`
components, and have them respond to changes in their props by auto
calling addListener or removeListener as needed.