Allow custom frontend entrypoints to be used
To allow for custom components to be fully end-to-end usable,
it needs to be possible for users of this library to load their frontend
code into the browser, and register their renderer with arcane js.

To facilitate this, when starting the Toolkit,
a custom entrypoint.js file path can be specified.
The `.map` file is also automatically made available.

The URL paths to load the file and the local file paths are determined
during startup, and a direct 1-1 mapping,
avoiding potential path traversal attacks.